machine-learning-client/web-app/requirements.txt
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.2.2
Flask-PyMongo==2.3.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
PocketSphinx==5.0.0
pydub==0.25.1
pymongo==4.2.0
python-dotenv==0.21.0
requests==2.28.1
SpeechRecognition==3.8.1
Werkzeug==2.2.2
zipp==3.9.0
wave==0.0.2
